backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pgvector>=0.3.6

# Validation & settings
pydantic>=2.10.0
pydantic-settings>=2.7.0
email-validator>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0

# Task queue
celery[redis]>=5.4.0
redis>=5.2.0

# HTTP client
httpx>=0.28.0

# AI / LLM
openai>=1.60.0

# Report generation
weasyprint>=63.0
jinja2>=3.1.4

# PII detection
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
httpx>=0.28.0
ruff>=0.8.0
mypy>=1.14.0
